rabbit review
