+++ app/src/common/utils/folderUtils/index.ts
@@ -18,6 +18,7 @@ export {
   getParentFoldersIds,
   getParentFolders,
   transformFoldersToDisplay,
+  collectAllTransformedFolderIds,
   getAllFolderIdsToDelete,
   getAllSubfolderIds,
   type BaseFolder,
