+++ app/src/pages/inside/common/expandedOptions/useSearchFilteredFolders.ts
@@ -14,38 +14,18 @@
  * limitations under the License.
  */
 
-import { useEffect, useMemo, useState, useCallback, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { isEmpty } from 'es-toolkit/compat';
 
 import {
   Folder,
-  TransformedFolder,
   foldersSelector,
   filteredFoldersSelector,
   isLoadingFilteredFoldersSelector,
   getFilteredFoldersAction,
   clearFilteredFoldersAction,
 } from 'controllers/testCase';
-import { transformFoldersToDisplay, getParentFoldersIds } from 'common/utils/folderUtils';
-
-const collectAllFolderIds = (folders: TransformedFolder[]): number[] => {
-  const ids: number[] = [];
-
-  const traverse = (list: TransformedFolder[]) => {
-    list.forEach((folder) => {
-      ids.push(folder.id);
-
-      if (!isEmpty(folder.folders)) {
-        traverse(folder.folders);
-      }
-    });
-  };
-
-  traverse(folders);
-
-  return ids;
-};
+import { useSearchFilteredFolderDisplay } from 'common/hooks';
 
 interface UseSearchFilteredFoldersParams {
   searchQuery?: string;
@@ -63,22 +43,11 @@ export const useSearchFilteredFolders = ({
   const allFolders = allFoldersOverride ?? defaultFolders;
   const filteredFolderData = useSelector(filteredFoldersSelector);
   const isLoading = useSelector(isLoadingFilteredFoldersSelector);
-  const [collapsedIds, setCollapsedIds] = useState<Set<number>>(new Set());
-  const prevSearchQueryRef = useRef(searchQuery);
   const initialFoldersRef = useRef(allFolders);
-  const respondedQueryRef = useRef<string | undefined>(undefined);
-
-  const isQueryPending = !!searchQuery && searchQuery !== respondedQueryRef.current;
 
   useEffect(() => {
-    if (prevSearchQueryRef.current !== searchQuery) {
-      setCollapsedIds(new Set());
-      prevSearchQueryRef.current = searchQuery;
-    }
-
     if (!searchQuery) {
       dispatch(clearFilteredFoldersAction());
-      respondedQueryRef.current = undefined;
       return;
     }
 
@@ -95,90 +64,10 @@ export const useSearchFilteredFolders = ({
     dispatch(getFilteredFoldersAction({ searchQuery, extraFilters }));
   }, [allFolders, searchQuery, extraFilters, dispatch]);
 
-  useEffect(() => {
-    if (!isLoading && searchQuery) {
-      respondedQueryRef.current = searchQuery;
-    }
-  }, [isLoading, searchQuery]);
-
-  const relevantFolderIds = useMemo(() => {
-    if (!searchQuery || filteredFolderData.length === 0) return new Set<number>();
-
-    const filteredIds = new Set(filteredFolderData.map((folder) => folder.id));
-    const idsWithAncestors = new Set<number>(filteredIds);
-
-    filteredIds.forEach((id) => {
-      getParentFoldersIds(id, allFolders).forEach((parentId) => idsWithAncestors.add(parentId));
-    });
-
-    return idsWithAncestors;
-  }, [filteredFolderData, allFolders, searchQuery]);
-
-  const filteredCountsMap = useMemo(
-    () => new Map(filteredFolderData.map((folder) => [folder.id, folder.countOfTestCases])),
-    [filteredFolderData],
-  );
-
-  const relevantFolders = useMemo(() => {
-    if (!searchQuery || relevantFolderIds.size === 0) return [];
-
-    return allFolders
-      .filter((folder) => relevantFolderIds.has(folder.id))
-      .map((folder) => {
-        const filteredCount = filteredCountsMap.get(folder.id);
-
-        return { ...folder, countOfTestCases: filteredCount ?? 0 };
-      });
-  }, [allFolders, relevantFolderIds, searchQuery, filteredCountsMap]);
-
-  const transformedFilteredFolders = useMemo(
-    () => (searchQuery ? transformFoldersToDisplay(relevantFolders) : []),
-    [relevantFolders, searchQuery],
-  );
-
-  const allFilteredFolderIds = useMemo(
-    () => (searchQuery ? collectAllFolderIds(transformedFilteredFolders) : []),
-    [transformedFilteredFolders, searchQuery],
-  );
-
-  const expandedFilteredFolderIds = useMemo(
-    () => allFilteredFolderIds.filter((id) => !collapsedIds.has(id)),
-    [allFilteredFolderIds, collapsedIds],
-  );
-
-  const handleToggleSearchFilteredFolder = useCallback(
-    (folder: TransformedFolder) => {
-      setCollapsedIds((prev) => {
-        const next = new Set(prev);
-
-        if (next.has(folder.id)) {
-          next.delete(folder.id);
-        } else {
-          next.add(folder.id);
-        }
-
-        return next;
-      });
-    },
-    [],
-  );
-
-  const hasFilteredFolders = !isEmpty(transformedFilteredFolders);
-
-  const filteredTotalTestCases = useMemo(() => {
-    if (!searchQuery) {
-      return 0;
-    }
-
-    return filteredFolderData.reduce((sum, folder) => sum + (folder.countOfTestCases || 0), 0);
-  }, [filteredFolderData, searchQuery]);
-
-  return {
-    searchFilteredFolders: transformedFilteredFolders,
-    searchFilteredExpandedIds: expandedFilteredFolderIds,
-    isSearchFilteredLoading: isLoading || isQueryPending,
-    hasSearchFilteredFolders: hasFilteredFolders,
-    handleToggleSearchFilteredFolder,
-    filteredTotalTestCases,
-  };
+  return useSearchFilteredFolderDisplay({
+    searchQuery,
+    allFolders,
+    filteredFolderData,
+    isLoading,
+  });
 };
