+++ app/src/pages/inside/manualLaunchesPage/messages.ts
@@ -24,6 +24,6 @@ export const messages = defineMessages({
   },
   searchPlaceholder: {
     id: 'ManualLaunchesPage.searchPlaceholder',
-    defaultMessage: 'Search launch by name',
+    defaultMessage: 'Search launches by name',
   },
 });
